driver andorcam2: don't warn about unsupported trigger mode if it's known
In case the camera doesn't support software trigger mode, we show a
warning at init.
By mistake, it would also show another warning every time a new
synchronized acquisition is started, which is confusing. That should
only be shown for the camera which do support software trigger, but a
setting is preventing its use.